webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   try {
@@ -11,8 +12,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const id = req.params.id;
-  const video = await Video.findById(id);
-  console.log(video.fileUrl);
+  const video = await Video.findById(id).populate("owner");
   try {
     return res.render("videos/watch", {
       pageTitle: video.title,
@@ -25,8 +25,14 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const id = req.params.id;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   try {
+    if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+    }
     return res.render("videos/edit", {
       pageTitle: `Edit : ${video.title} `,
       video,
@@ -39,16 +45,19 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  try {
-    await Video.findByIdAndUpdate(id, {
-      title,
-      description,
-      hashtags: Video.formatHashtags(hashtags),
-    });
-    return res.redirect(`/`);
-  } catch {
-    return res.status(404).render("404", { pageTitle: "Video Not Found" });
+  const video = await Video.exists({ _id: id });
+  const videoModified = await Video.findByIdAndUpdate(id, {
+    title,
+    description,
+    hashtags: Video.formatHashtags(hashtags),
+  });
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
   }
+  if (String(videoModified.owner) !== String(video._id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.redirect(`/videos/${id}`);
 };
 
 export const getUpload = (req, res) => {
@@ -56,15 +65,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { path } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl: path ? path : fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     console.log(error);
@@ -77,12 +93,22 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
-  try {
-    await Video.findByIdAndDelete(id);
-    return res.redirect("/");
-  } catch {
-    return res.status(404).render("404", { pageTitle: "Video Not Found" });
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
+  await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
+  return res.redirect("/");
 };
 
 //Global Router
